Fix FusedLinearJSD precision issue when using AMP (#336)
## Summary
<!--- This is a required section; please describe the main purpose of
this proposed code change. --->
1. make sure all the computation between logit to final JSD loss happen
on FP32
2. make sure FLJSD works properly under mixed precision scenario, also
add a test to guard
3. the Torch CE loss impl we use in testing FLCE misses out the fp32
cast for logits, add it back. **NOTE: we should definitely jus switch
directly to [HF
impl](https://github.com/huggingface/transformers/blob/main/src/transformers/loss/loss_utils.py#L32)
for testing to ensure always doing apple-to-apple comparison. See the
added TODO item.**


<!---
## Details
This is an optional section; is there anything specific that reviewers
should be aware of?
--->

## Testing Done
<!--- This is a required section; please describe how this change was
tested. --->

<!-- 
Replace BLANK with your device type. For example, A100-80G-PCIe

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them. 
-->

- Hardware Type: <BLANK>
- [x] run `make test` to ensure correctness
- [x] run `make checkstyle` to ensure code style
- [x] run `make test-convergence` to ensure convergence

+++ test/transformers/test_fused_linear_cross_entropy.py
@@ -22,6 +22,10 @@ class TorchLMHeadCE(torch.nn.Module):
     :param V: vocab size
     :param ignore_index: index to ignore
     :param reduction: reduction method
+
+    # TODO: if we bump CI env's `transformers` version to >= 4.46, we should just directly
+    # call https://github.com/huggingface/transformers/blob/main/src/transformers/loss/loss_utils.py#L32
+    # to be consistent with Hugging Face model implementation.
     """
 
     def __init__(
@@ -45,7 +49,7 @@ class TorchLMHeadCE(torch.nn.Module):
         )
 
     def forward(self, x, y):
-        logits = self.lin(x)
+        logits = self.lin(x).to(torch.float32)
         return self.ce_loss(logits, y)
